[CBRD-27365] PR-2a: classify backward-capable lists (QFILE_FLAG_BACKWARD) and assert the hdr_size contracts
Records the #184 classification at every qfile_open_list site without changing behavior: PR-2a keeps
hdr_size 8 for every list, PR-2b derives hdr_size from the flag (ADR 0016 D-181-8).

- (A) top-most XASL results: ORDER BY/DISTINCT sort output, GROUP BY (sort and index), merge join,
  BUILDLIST/BUILD_SCHEMA/BUILDVALUE, UNION/DIFFERENCE/INTERSECTION, CTE, analytic output, hash join
  (manager->qlist_flag), px writer lists -> XASL_LIST_BACKWARD_FLAG (xasl).
- (B) MERGELIST_PROC children: new XASL_LIST_BACKWARD flag set by ptqo_to_merge_list_proc, folded into
  the same helper.
- (C) analytic group/value lists (4 open sites): QFILE_FLAG_BACKWARD unconditionally.
- Contracts (trivially true now, bite in PR-2b): qfile_scan_prev / cursor_prev_tuple assert the list is
  backward capable; qfile_connect_list / qfile_append_list assert equal hdr_size; the child-to-result
  promotions (CTE recursive union, hash join, px merged result) assert a top-most result is backward capable.

+++ src/query/qfile_tuple_layout.c
@@ -297,7 +297,7 @@ qfile_slot_overwrite_value (QFILE_TUPLE_RECORD * rec, int col, const TP_DOMAIN *
 
   if (value == NULL || DB_IS_NULL (value) || is_null || dom == NULL || dom->type == NULL
       || TP_DOMAIN_TYPE (dom) != DB_VALUE_DOMAIN_TYPE (value)
-      || QFILE_LEGACY_VALUE_ENCODED_SIZE (pr_data_writeval_disk_size ((DB_VALUE *) value)) != len)
+      || (int) QFILE_LEGACY_VALUE_ENCODED_SIZE (pr_data_writeval_disk_size ((DB_VALUE *) value)) != len)
     {
       assert (false);
       return ER_FAILED;
